feat: use constant-time Base64 encoding
We already have the `base64ct` crate in the repo (dep of
ethers I believe) and I think one is enough! It is also
constant time, which is strictly better than the current
`base64` and seem to have a better API.

feat: cleaning up the synchronizer (#306)

feat: use constant-time Base64 encoding

We already have the `base64ct` crate in the repo (dep of ethers I believe) and I think one is enough! It is also constant time, which is strictly better than the current `base64` and seem to have a better API.

feat: use constant-time Base64 encoding

We already have the `base64ct` crate in the repo (dep of
ethers I believe) and I think one is enough! It is also
constant time, which is strictly better than the current
`base64` and seem to have a better API.

feat: cleaning up the synchronizer (#306)

feat: use constant-time Base64 encoding

We already have the `base64ct` crate in the repo (dep of ethers I believe) and I think one is enough! It is also constant time, which is strictly better than the current `base64` and seem to have a better API.